field names of changed models
