+++ backend/src/routes/index.js
@@ -1,8 +1,8 @@
 const {Router} = require("express");
 const router = Router();
 const comments = require('../models/comments')
-const publications = require('../models/publicacion')
-const users = require('../models/users')
+const publications = require('../models/publication')
+const users = require('../models/user')
